+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- Android: `onEvent` callback not being called on `PlayerView`
+
 ## [0.14.1] (2023-11-16)
 
 ### Fixed