+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixed
 
-- iOS: `onEvent` on iOS has partial payload information
+- iOS: `onEvent` on iOS has incomplete payload information
 
 ## [0.14.1] (2023-11-16)