+++ CHANGELOG.md
@@ -6,13 +6,13 @@
 
 - Android: `onEvent` callback not being called on `PlayerView`
 - iOS: `onEvent` on iOS has incomplete payload information
+- tvOS: Picture in Picture sample screen has unwanted padding
 
 ## [0.14.1] (2023-11-16)
 
 ### Fixed
 
 - Android: `PlayerView` destroys attached `Player` instance on destroy. `Player` lifecycle must be handled on the creation side
-- tvOS: Picture in Picture sample screen has unwanted padding
 
 ## [0.14.0] (2023-11-14)